feat: beacon_api test crate
