+++ Cargo.toml
@@ -16,6 +16,7 @@ members = [
     "crates/rpc",
     "crates/runtime",
     "crates/storage",
+    "testing/beacon-api",
     "testing/ef-tests",
 ]
 resolver = "2"
