+++ testing/beacon-api/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "beacon-api-tests"
+authors.workspace = true
+edition.workspace = true
+keywords.workspace = true
+license.workspace = true
+readme.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+alloy-primitives.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+tokio.workspace = true
+
+# ream
+ream-consensus.workspace = true
+ream-rpc.workspace = true
