+++ testing/ef-tests/Cargo.toml
@@ -18,7 +18,6 @@ ethereum_ssz_derive.workspace = true
 paste = "1.0.15"
 rstest.workspace = true
 serde.workspace = true
-serde_json.workspace = true
 serde_yaml.workspace = true
 snap.workspace = true
 ssz_types.workspace = true
@@ -29,7 +28,6 @@ tree_hash.workspace = true
 ream-consensus.workspace = true
 ream-fork-choice.workspace = true
 ream-merkle.workspace = true
-ream-rpc.workspace = true
 
 [features]
 ef-tests = []
